moved worker entrypoints, added a readme to help explain this
