+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './exports/browser'
 export * from './exports/durable_object'
-export { sharedWorkerEntrypoint } from './exports/shared_worker'
-export * from './exports/worker'
+export { sharedWorkerEntrypoint } from './exports/entrypoints/shared_worker'
+export * from './exports/entrypoints/worker'
 export type { Transition } from './types/transitions/Transition'
